fix: split pair_accepted tests to satisfy rust structural file-size gate
