+++ crates/clawdentity-cli/src/commands/connector/AGENTS.md
@@ -26,6 +26,7 @@
 - Prefer `&Path` in internal helper signatures and only use `PathBuf` where ownership is required, so Clippy `ptr_arg` remains green in connector runtime code.
 - Handle pairing acceptance system events in `delivery/pair_accepted.rs` and invoke that processor in both live inbound delivery flow and retry replay flow.
 - Keep pair-accepted peer persistence idempotent by reusing core helper `persist_confirmed_peer_from_profile_and_proxy_origin`; registry enrichment from `GET /v1/agents/profile` is best-effort and must not block trusted peer persistence when registry is unavailable.
+- Keep `delivery/pair_accepted.rs` under structural line limits by moving larger test coverage into `delivery/pair_accepted/tests.rs` rather than growing the runtime module file.
 - Pair-accepted system side effects must run only for trusted relay delivery provenance (`deliverySource=proxy.events.queue.pair_accepted`); never mutate peer state for user-authored payload-only `system.type=pair.accepted`.
 - Pair-accepted system payload validation must include DID checks, responder proxy origin URL checks, and event timestamp parsing before mutating peer state.
 - Pair-accepted structured fields are mandatory for trusted side effects; optional `system.message` is UX-only and must never be used as a replacement for persistence/trust metadata.
